新增google authenticator验证并配置开关

+++ ruoyi/src/main/java/com/ruoyi/framework/security/service/SysLoginService.java
@@ -1,6 +1,11 @@
 package com.ruoyi.framework.security.service;
 
 import javax.annotation.Resource;
+
+import com.ruoyi.common.exception.user.GoogleAuthCodeException;
+import com.ruoyi.common.utils.GoogleAuthenticator;
+import com.ruoyi.common.utils.StringUtils;
+import com.ruoyi.project.system.domain.SysUser;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.BadCredentialsException;
@@ -44,20 +49,22 @@ public class SysLoginService
      * @param uuid 唯一标识
      * @return 结果
      */
-    public String login(String username, String password, String code, String uuid)
+    public String login(String username, String password, String code, String uuid, String googlecode)
     {
-        String verifyKey = Constants.CAPTCHA_CODE_KEY + uuid;
-        String captcha = redisCache.getCacheObject(verifyKey);
-        redisCache.deleteObject(verifyKey);
-        if (captcha == null)
-        {
-            AsyncManager.me().execute(AsyncFactory.recordLogininfor(username, Constants.LOGIN_FAIL, MessageUtils.message("user.jcaptcha.expire")));
-            throw new CaptchaExpireException();
-        }
-        if (!code.equalsIgnoreCase(captcha))
-        {
-            AsyncManager.me().execute(AsyncFactory.recordLogininfor(username, Constants.LOGIN_FAIL, MessageUtils.message("user.jcaptcha.error")));
-            throw new CaptchaException();
+        if (tokenService.isCaptchaEnabled()) {
+            String verifyKey = Constants.CAPTCHA_CODE_KEY + uuid;
+            String captcha = redisCache.getCacheObject(verifyKey);
+            redisCache.deleteObject(verifyKey);
+            if (captcha == null)
+            {
+                AsyncManager.me().execute(AsyncFactory.recordLogininfor(username, Constants.LOGIN_FAIL, MessageUtils.message("user.jcaptcha.expire")));
+                throw new CaptchaExpireException();
+            }
+            if (!code.equalsIgnoreCase(captcha))
+            {
+                AsyncManager.me().execute(AsyncFactory.recordLogininfor(username, Constants.LOGIN_FAIL, MessageUtils.message("user.jcaptcha.error")));
+                throw new CaptchaException();
+            }
         }
         // 用户验证
         Authentication authentication = null;
@@ -66,6 +73,19 @@ public class SysLoginService
             // 该方法会去调用UserDetailsServiceImpl.loadUserByUsername
             authentication = authenticationManager
                     .authenticate(new UsernamePasswordAuthenticationToken(username, password));
+            if (tokenService.isGoogleAuthenticator()) {
+                LoginUser principal = (LoginUser) authentication.getPrincipal();
+                SysUser user = principal.getUser();
+                String googlekey = user.getGooglekey();
+                if (StringUtils.isNotEmpty(googlecode)) {
+                    String totpCode = GoogleAuthenticator.getTOTPCode(googlekey);
+                    if (!googlecode.equals(totpCode)) {
+                        throw new GoogleAuthCodeException();
+                    }
+                } else {
+                    throw new GoogleAuthCodeException();
+                }
+            }
         }
         catch (Exception e)
         {
